Forgot to override before delete

+++ models/Customer.php
@@ -69,7 +69,9 @@ class Customer extends \yii\db\ActiveRecord
     public function beforeDelete()
     {
         $phone = $this->phone;
-        $phone->delete();
+        if(!empty($phone))
+            $phone->delete();
+        return parent::beforeDelete();
         
     }
     /**
